secure_file_system/requirements/dev.txt
-r base.txt
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.2.1
coverage==7.2.7
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.8.0
# Type stubs
django-stubs==4.2.5
djangorestframework-stubs==3.14.0
types-python-dateutil==2.9.0.20240316
types-pytz>=2024.1.0
typing-extensions>=4.5.0  # Required by django-stubs
